ajout gestion site en fonction du role

+++ app/Http/Controllers/PartitionController.php
@@ -13,11 +13,21 @@ class PartitionController extends Controller
      */
     public function index()
     {
+        $partitions = Partition::latest()->paginate(12);
+
         return view('partitions.index', compact('partitions'));
+    }
+
+    public function show(Partition $partition)
+    {
+        $this->authorize('view', $partition);
+
+        return view('partitions.show', compact('partition'));
     }
 
     public function create()
     {
+        $this->authorize('create', Partition::class);
 
         return view('partitions.create');
     }
@@ -27,16 +37,22 @@ class PartitionController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Partition::class);
 
         $validated = $request->validate([
+            'title' => ['required', 'string', 'max:255'],
+            'composer' => ['nullable', 'string', 'max:255'],
         ]);
 
         // Upload du fichier MusicXML
         $path = $request->file('musicxml_file')->store('partitions', 'public');
 
         $partition = Partition::create([
+            ...$validated,
+            'user_id' => $request->user()->id,
         ]);
 
+        return redirect()->route('partitions.show', $partition);
     }
 
     /**
@@ -44,6 +60,7 @@ class PartitionController extends Controller
      */
     public function edit(Partition $partition)
     {
+        $this->authorize('update', $partition);
 
         return view('partitions.edit', compact('partition'));
     }
@@ -53,11 +70,16 @@ class PartitionController extends Controller
      */
     public function update(Request $request, Partition $partition)
     {
+        $this->authorize('update', $partition);
 
         $validated = $request->validate([
+            'title' => ['required', 'string', 'max:255'],
+            'composer' => ['nullable', 'string', 'max:255'],
         ]);
 
+        $partition->update($validated);
 
+        return redirect()->route('partitions.show', $partition);
     }
 
         $partition->save();
@@ -71,9 +93,11 @@ class PartitionController extends Controller
      */
     public function destroy(Partition $partition)
     {
+        $this->authorize('delete', $partition);
 
         $partition->delete();
 
+        return redirect()->route('partitions.index');
     }
 }
